Added password change to authentication
