+++ authentication/views.py
@@ -1,16 +1,31 @@
 import logging
-import jwt
-from authentication.models import CustomUser
+
 from authentication.permissions import CustomUserUpdatePermission
 from authentication.serializers import UserRegistrationSerializer, UserUpdateSerializer, UserPasswordUpdateSerializer
 from django.contrib.auth import authenticate
 from django.conf import settings
-from django.http import JsonResponse
+
 from rest_framework import status, generics
-from rest_framework.views import APIView
-from rest_framework.response import Response
+
 from rest_framework_simplejwt.tokens import RefreshToken
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser
+import datetime
+
+import jwt
+from django.contrib.auth.hashers import make_password
+from django.core.exceptions import ValidationError
+
+from django.core.validators import validate_email
+from django.http import JsonResponse
+from django.shortcuts import render
+
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from authentication.models import CustomUser
+from forum import settings
+from .serializers import PasswordRecoverySerializer
+from .utils import Utils
 
 
 class UserRegistrationView(generics.CreateAPIView):
@@ -77,3 +92,52 @@ class LogoutView(APIView):
             print("Exception", e)
             return Response(status=status.HTTP_400_BAD_REQUEST)
 
+
+class PasswordRecoveryAPIView(APIView):
+    def post(self, request):
+        email = request.data.get('email')
+        try:
+            validate_email(email)
+        except ValidationError:
+            return Response({'error': 'Invalid email address'}, status=400)
+
+        try:
+            user = CustomUser.objects.get(email=email)
+        except CustomUser.DoesNotExist:
+            return Response({'error': 'User with this email does not exist'}, status=404)
+
+        jwt_token = Utils.generate_token(user.email, user.pk)
+        reset_link = f"{settings.FRONTEND_URL}/auth/password-reset/{jwt_token}/"
+        try:
+            Utils.send_password_reset_email(email, reset_link)
+        except Exception as e:
+            return Response({'error': 'Failed to send email'}, status=500)
+        return Response({'message': 'Password reset email sent successfully'})
+
+
+
+class PasswordResetView(APIView):
+    serializer_class = PasswordRecoverySerializer
+
+    def get(self, request, jwt_token):
+
+        uid, email, exp = Utils.decode_token(jwt_token)
+        user = Utils.get_user(uid, email)
+
+        if user is not None:
+            return render(request, 'reset_password_form.html', {'jwt_token': jwt_token})
+        else:
+            return JsonResponse({'error': 'Invalid token for password reset'}, status=400)
+
+    def post(self, request, jwt_token):
+
+        uid, email, exp = Utils.decode_token(jwt_token)
+
+        user = Utils.get_user(uid, email)
+        if user is not None:
+            serializer = self.serializer_class(instance=user, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return render(request, 'password_reset_done.html')
+        else:
+            return JsonResponse({'error': 'Invalid token for password reset'}, status=status.HTTP_400_BAD_REQUEST)
