+++ forum/settings.py
@@ -47,7 +47,7 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt.token_blacklist',
     'rest_framework',
     'authentication',
-    'password_recovery'
+
 ]
 
 REST_FRAMEWORK = {
