+++ forum/urls.py
@@ -7,7 +7,7 @@ from rest_framework_simplejwt.views import  TokenRefreshView
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('auth/', include('authentication.urls')),
-    path('recovery/', include('password_recovery.urls')),
+    # path('recovery/', include('password_recovery.urls')),
     path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
     path("api/login", LoginView.as_view(), name='login'),
