--- password_recovery/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class PasswordRecoveryConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'password_recovery'
